feat: 메인 카드 섹션 추가
- 메인 카드 섹션을 추가합니다.

#3
